feat: ForwardDiff autodiff, thread-parallel restarts, 100% coverage
- QAOAAngles{T<:Real}: parametric struct enables ForwardDiff dual number
  propagation through the entire Basso evaluation pipeline
- Generalized ~55 type barriers in basso_finite_d.jl: Float64 → T,
  ComplexF64 → Complex{T}, removed concrete return annotations
- optimizer uses autodiff=AutoForwardDiff() for exact analytic gradients
  (eliminates ~2p+1 FD evaluations per gradient step)
- Thread-parallel restarts via Threads.@threads in optimize_angles
- Added ForwardDiff and ADTypes as direct dependencies
- New tests: Tier 1 state-vector internals, merge_optimization_results
  edge cases, Tier 1 vs Tier 2 cross-validation
- 690 tests passing, 672/672 lines covered (100%)
- Spec: .project/specs/autodiff-generics.md

+++ src/qaoa.jl
@@ -165,7 +165,7 @@ function reference_parity_expectation(
     params::TreeParams,
     angles::QAOAAngles;
     clause_sign::Int=1,
-)::Float64
+)
     tree, state = simulate_light_cone_state(params, angles; clause_sign)
     root_parity_expectation(state, tree.root_clause)
 end
@@ -180,7 +180,7 @@ function parity_expectation(
     params::TreeParams,
     angles::QAOAAngles;
     clause_sign::Int=1,
-)::Float64
+)
     basso_parity_expectation(params, angles; clause_sign)
 end
 
@@ -200,6 +200,6 @@ function qaoa_expectation(
     params::TreeParams,
     angles::QAOAAngles;
     clause_sign::Int=1,
-)::Float64
+)
     basso_expectation(params, angles; clause_sign)
 end
